+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Feature - Adding logging to event handler startup
 ## Current Release 
 ### 0.224.0 
 **Release Date:** Thu Sep 27 17:41:43 UTC 2018